Extract bootstrap_context from bootstrap_service
Strategies now depend on bootstrap_context instead of the full service.
