+++ nano/lib/thread_roles.cpp
@@ -121,6 +121,9 @@ std::string nano::thread_role::get_string (nano::thread_role::name role)
 		case nano::thread_role::name::bootstrap_dependency_walker:
 			thread_role_name_string = "Bootstrap walkr";
 			break;
+		case nano::thread_role::name::bootstrap_dependency_sync:
+			thread_role_name_string = "Bootstrap dsync";
+			break;
 		case nano::thread_role::name::bootstrap_frontier_scan:
 			thread_role_name_string = "Bootstrap front";
 			break;
