+++ nano/node/bootstrap/database_strategy.cpp
@@ -5,8 +5,8 @@
 
 namespace nano::bootstrap
 {
-database_strategy::database_strategy (nano::bootstrap_service & service_a) :
-	service{ service_a }
+database_strategy::database_strategy (bootstrap_context & ctx_a) :
+	ctx{ ctx_a }
 {
 }
 
@@ -26,13 +26,13 @@ void database_strategy::stop ()
 
 void database_strategy::run ()
 {
-	nano::unique_lock<nano::mutex> lock{ service.mutex };
-	while (!service.stopped)
+	nano::unique_lock<nano::mutex> lock{ ctx.mutex };
+	while (!ctx.stopped)
 	{
 		// Avoid high churn rate of database requests
-		bool should_throttle = !service.database_scan.warmed_up () && service.throttle.throttled ();
+		bool should_throttle = !ctx.database_scan.warmed_up () && ctx.throttle.throttled ();
 		lock.unlock ();
-		service.stats.inc (nano::stat::type::bootstrap, nano::stat::detail::loop_database);
+		ctx.stats.inc (nano::stat::type::bootstrap, nano::stat::detail::loop_database);
 		run_one (should_throttle);
 		lock.lock ();
 	}
@@ -40,8 +40,8 @@ void database_strategy::run ()
 
 void database_strategy::run_one (bool should_throttle)
 {
-	service.wait_block_processor ();
-	auto channel = service.wait_channel ();
+	ctx.wait_block_processor ();
+	auto channel = ctx.wait_channel ();
 	if (!channel)
 	{
 		return;
@@ -51,35 +51,35 @@ void database_strategy::run_one (bool should_throttle)
 	{
 		return;
 	}
-	service.request (account, 2, channel, nano::bootstrap::query_source::database);
+	ctx.request (account, 2, channel, query_source::database);
 }
 
 nano::account database_strategy::next_database (bool should_throttle)
 {
-	debug_assert (!service.mutex.try_lock ());
-	debug_assert (service.config.database_warmup_ratio > 0);
+	debug_assert (!ctx.mutex.try_lock ());
+	debug_assert (ctx.config.database_warmup_ratio > 0);
 
 	// Throttling increases the weight of database requests
-	if (!service.database_limiter.should_pass (should_throttle ? service.config.database_warmup_ratio : 1))
+	if (!ctx.database_limiter.should_pass (should_throttle ? ctx.config.database_warmup_ratio : 1))
 	{
 		return { 0 };
 	}
-	auto account = service.database_scan.next ([this] (nano::account const & account) {
-		return service.count_tags (account, nano::bootstrap::query_source::database) == 0;
+	auto account = ctx.database_scan.next ([this] (nano::account const & account) {
+		return ctx.count_tags (account, query_source::database) == 0;
 	});
 	if (account.is_zero ())
 	{
 		return { 0 };
 	}
-	service.stats.inc (nano::stat::type::bootstrap_next, nano::stat::detail::next_database);
+	ctx.stats.inc (nano::stat::type::bootstrap_next, nano::stat::detail::next_database);
 	return account;
 }
 
 nano::account database_strategy::wait_database (bool should_throttle)
 {
 	nano::account result{ 0 };
-	service.wait ([this, &result, should_throttle] () {
-		debug_assert (!service.mutex.try_lock ());
+	ctx.wait ([this, &result, should_throttle] () {
+		debug_assert (!ctx.mutex.try_lock ());
 		result = next_database (should_throttle);
 		if (!result.is_zero ())
 		{
