+++ nano/node/bootstrap/dependency_strategy.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <nano/node/bootstrap/bootstrap_service.hpp>
+#include <nano/node/bootstrap/bootstrap_context.hpp>
 
 #include <thread>
 
@@ -9,22 +9,24 @@ namespace nano::bootstrap
 class dependency_strategy
 {
 public:
-	explicit dependency_strategy (nano::bootstrap_service & service);
+	explicit dependency_strategy (bootstrap_context & ctx);
 
 	void start ();
 	void stop ();
 	void run_one ();
 
-	bool process (nano::messages::asc_pull_ack::account_info_payload const & response, nano::bootstrap_service::async_tag const & tag);
+	bool process (nano::messages::asc_pull_ack::account_info_payload const & response, async_tag const & tag);
 
 private:
 	void run ();
+	void run_sync ();
 
 	nano::block_hash next_blocking ();
 	nano::block_hash wait_blocking ();
 	bool request_info (nano::block_hash hash, std::shared_ptr<nano::transport::channel> const & channel);
 
-	nano::bootstrap_service & service;
+	bootstrap_context & ctx;
 	std::thread thread;
+	std::thread sync_thread;
 };
 }
