+++ nano/node/bootstrap/priority_strategy.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <nano/node/bootstrap/bootstrap_service.hpp>
+#include <nano/node/bootstrap/bootstrap_context.hpp>
 
 #include <thread>
 
@@ -9,7 +9,7 @@ namespace nano::bootstrap
 class priority_strategy
 {
 public:
-	explicit priority_strategy (nano::bootstrap_service & service);
+	explicit priority_strategy (bootstrap_context & ctx);
 
 	void start ();
 	void stop ();
@@ -23,7 +23,7 @@ private:
 	priority_result next_priority ();
 	priority_result wait_priority ();
 
-	nano::bootstrap_service & service;
+	bootstrap_context & ctx;
 	std::thread thread;
 };
 }
